requirements.txt
Flask
requests
beautifulsoup4
google-generativeai
yt-dlp
python-dotenv